restyle top and bottom navs using flexbox
