+++ public/css/main.css
@@ -1,1 +1,1 @@
-*{margin:0;padding:0;box-sizing:border-box}button,input{-webkit-appearance:none}body{background-color:#020304}.clearfix{overflow:auto;zoom:1}footer{background-color:#541f14;color:#938172;display:block;font-size:75%;letter-spacing:.05em;margin-top:2em;padding:1em}html{font-size:62.5%}body{color:#fff;font-family:"Roboto",sans-serif;font-size:1.6rem;line-height:1.25em;text-align:center}h2{font-weight:300;margin:1em}input{border:2px solid #938172;border-radius:5px;background:none;color:#cc9e61;display:block;margin:1em auto;padding:.5em;letter-spacing:.2em;text-align:center;text-transform:uppercase;width:90%}input[type=submit]{cursor:pointer}input[type=submit]:hover{background-color:#938172;color:#020304}.top-nav{background-color:#938172;margin-bottom:6.25rem}.trailericon{display:block;margin:1rem auto;width:25%}.top-nav_links{float:right;position:relative}.top-nav_links a{color:#fff;display:inline-block;padding:1.5rem 1.5rem 1.5rem 0;text-decoration:none}.top-nav_links a:hover{color:#541F14;font-family:airstream;font-size:1.5em;padding:1.21875rem 1.21875rem 1.21875rem 0}
+*{margin:0;padding:0;box-sizing:border-box}button,input{-webkit-appearance:none}body{background-color:#020304}.clearfix{overflow:auto;zoom:1}.app-container{display:flex;flex-direction:column;height:100vh}main{flex-grow:1;flex-shrink:1;flex:1;overflow:auto;margin:0 auto;max-width:960px;width:100%}@media (min-width: 480px){main{overflow:visible}}html{font-size:62.5%}body{color:#fff;font-family:"Roboto",sans-serif;font-size:1.6rem;line-height:1.25em;text-align:center}h2{font-weight:300;margin:1em}input{border:2px solid #938172;border-radius:5px;background:none;color:#cc9e61;display:block;margin:1em auto;padding:.5em;letter-spacing:.2em;text-align:center;text-transform:uppercase;width:90%}input[type=submit]{cursor:pointer}input[type=submit]:hover{background-color:#938172;color:#020304}nav{display:flex;align-items:center;justify-content:space-between}nav a{color:#fff;display:inline-block;text-decoration:none}nav a:hover{color:#e6e6e6}header{background-color:#938172;display:flex;align-items:center;justify-content:center;flex-shrink:0;padding:1rem}header .logo{flex:1;max-width:7.5rem;width:25%}header .image-container{padding-bottom:70.8%;height:0;position:relative}header .logo img{width:100%;height:100%;display:block;position:absolute}header nav{display:none;flex:2;justify-content:flex-end}header nav a{font-size:87.5%;padding:1em}@media (min-width: 480px){header .logo img{float:left}header nav{display:flex}}footer{background-color:#541f14;color:#938172;display:block;flex-shrink:0;font-size:75%;padding:1em 2em}footer p{display:none}@media (min-width: 480px){footer nav{display:none}footer p{display:block}}
