+++ sass/partials/components/_all.scss
@@ -1,1 +1,3 @@
-@import 'top-nav';
+@import 'nav';
+@import 'header';
+@import 'footer';
