+++ sass/partials/components/_footer.scss
@@ -1,0 +1,22 @@
+footer {
+  background-color: $maroon;
+  color: $taupe;
+  display: block;
+  flex-shrink: 0;
+  font-size: 75%;
+  padding: 1em 2em;
+}
+
+footer p {
+  display: none;
+}
+
+@media (min-width: 480px) {
+  footer nav {
+    display: none;
+  }
+
+  footer p {
+    display: block;
+  }
+}
