+++ sass/partials/components/_header.scss
@@ -1,0 +1,49 @@
+header {
+  background-color: $taupe;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-shrink: 0;
+  padding: 1rem;
+}
+
+header .logo {
+  flex: 1;
+  max-width: 7.5rem;
+  width: 25%;
+}
+
+header .image-container {
+  padding-bottom: 70.8%;
+  height: 0;
+  position: relative;
+}
+
+header .logo img {
+  width: 100%;
+  height: 100%;
+  display: block;
+  position: absolute;
+}
+
+header nav {
+  display: none;
+  flex: 2;
+  justify-content: flex-end;
+}
+
+header nav a {
+  font-size: 87.5%;
+  padding: 1em;
+}
+
+@media (min-width: 480px) {
+  header .logo img {
+    float: left;
+  }
+
+  header nav {
+    display: flex;
+  }
+
+}
