--- sass/partials/components/_top-nav.scss
@@ -1,27 +1,0 @@
-.top-nav {
-  background-color: $taupe;
-  margin-bottom: 6.25rem;
-}
-
-.trailericon {
-  display: block;
-  margin: 1rem auto;
-  width: 25%;
-}
-
-.top-nav_links {
-  float: right;
-  position:  relative;
-}
-.top-nav_links a {
-  color: #fff;
-  display: inline-block;
-  padding: 1.5rem 1.5rem 1.5rem 0;
-  text-decoration: none;
-}
-.top-nav_links a:hover {
-  color: #541F14;
-  font-family: airstream;
-  font-size: 1.5em;
-  padding: 1.21875rem 1.21875rem 1.21875rem 0;
-}
